requirements.txt
# Core dependencies (existing)
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# ML Forecasting dependencies (new)
statsmodels==0.14.0
tensorflow==2.13.0
keras==2.13.1
prophet==1.1.4
matplotlib==3.7.2
seaborn==0.12.2

# Jupyter notebook dependencies
jupyter==1.0.0
notebook==7.0.0
ipykernel==6.25.0
ipywidgets==8.1.0

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
